Allow user to get the list of extra fields present in the entry.
There are well known fields in the entry that are always available.
Other fields are also available but there is no way to know what they are
but trying to retrieve them.

The method get_extra_fields allow user code to get the list (iterator) of
extra fields present in the entry.