Return a new entry instead of taking an entry as input to change it.
Taking a argument as ouput is a C idiom. In python, we return the value.

The entry is a plain dict, not anymore a class with Mapping API.

More than the pythonic API, this fix also a bug :
The C tagEntry share the same internal buffer to store the datas.
ctags* function "just" changes the entries' pointers to the right place
in the buffer. Using one of those functions will invalidate any other
entries already returned. By using a dict we make a plain copy of the
string and so entry are now independent of the internal state of the C
code.